+++ backend/app/Http/Middleware/Term/ValidateTermEditPermission.php
@@ -19,7 +19,7 @@ class ValidateTermEditPermission
         $study_set_id = (int)$request->route()->parameter('study_set_id');
         $term_id = (int)$request->route()->parameter('term_id');
 
-        if (TermPermissionService::getParentStudySetId($term_id) !== $study_set_id) {
+        if ((int)TermPermissionService::getParentStudySetId($term_id) !== $study_set_id) {
             throw new RestResourceNotFoundException('Requested pair of study set and term does not exist.');
         }
